refactor : model 패키지로 이동
